Move back reporting engine code now that separate implementations are not needed
