+++ src/app/BUILD.gn
@@ -199,8 +199,6 @@ static_library("interaction-model") {
     #
     #       This breaks having `.h` and `.cpp` based off the same compilation unit and
     #       should ideally be cleaned up.
-    "reporting/Read.cpp",
-    "reporting/Read.h",
     "reporting/reporting.h",
   ]
